# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HYDRA-UMC-SERVER API",
    "version": "0.7.8",
    "description": "Generated from the routes registered in src/. It lists paths, methods, required access, the status codes each handler can return, and the names of the JSON body fields and query parameters it reads. Value types and the response bodies are described in docs/REMOTE_API.md."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/system/bluetooth/pair": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "POST /api/system/bluetooth/pair",
        "description": "Requires a signed-in account with the admin role.",
        "x-access": "admin",
        "responses": {
          "200": {
            "description": "Success. The body is documented in docs/REMOTE_API.md."
          },
          "400": {
            "description": "Invalid request. The body is documented in docs/REMOTE_API.md."
          },
          "503": {
            "description": "Not configured or unavailable. The body is documented in docs/REMOTE_API.md."
          }
        },
        "requestBody": {
          "required": false,
          "description": "JSON body. Only the field names the handler reads are listed; types and limits are checked in the handler.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mac": {}
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/system/bluetooth/power": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "POST /api/system/bluetooth/power",
        "description": "Requires a signed-in account with the admin role.",
        "x-access": "admin",
        "responses": {
          "200": {
            "description": "Success. The body is documented in docs/REMOTE_API.md."
          },
          "400": {
            "description": "Invalid request. The body is documented in docs/REMOTE_API.md."
          },
          "503": {
            "description": "Not configured or unavailable. The body is documented in docs/REMOTE_API.md."
          }
        },
        "requestBody": {
          "required": false,
          "description": "JSON body. Only the field names the handler reads are listed; types and limits are checked in the handler.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "on": {}
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/system/bluetooth/remove": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "POST /api/system/bluetooth/remove",
        "description": "Requires a signed-in account with the admin role.",
        "x-access": "admin",
        "responses": {
          "200": {
            "description": "Success. The body is documented in docs/REMOTE_API.md."
          },
          "400": {
            "description": "Invalid request. The body is documented in docs/REMOTE_API.md."
          },
          "503": {
            "description": "Not configured or unavailable. The body is documented in docs/REMOTE_API.md."
          }
        },
        "requestBody": {
          "required": false,
          "description": "JSON body. Only the field names the handler reads are listed; types and limits are checked in the handler.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mac": {}
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/system/bluetooth/scan": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "POST /api/system/bluetooth/scan",
        "description": "Requires a signed-in account with the admin role.",
        "x-access": "admin",
        "responses": {
          "200": {
            "description": "Success. The body is documented in docs/REMOTE_API.md."
          },
          "503": {
            "description": "Not configured or unavailable. The body is documented in docs/REMOTE_API.md."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/system/bluetooth/status": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "GET /api/system/bluetooth/status",
        "description": "Requires a signed-in account with the admin role.",
        "x-access": "admin",
        "responses": {
          "200": {
            "description": "Success. The body is documented in docs/REMOTE_API.md."
          },
          "503": {
            "description": "Not configured or unavailable. The body is documented in docs/REMOTE_API.md."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}